requirements.txt
django==4.1.7
django-imagekit==4.1.0
django-hashid-field==3.3.7
django-storages==1.13.2
Pillow==9.4.0
gunicorn==20.1.0
django-ninja==0.20.0
django-unicorn==0.50.0